route fixes, acars fixes, drop airways/navdb table before updating
git-svn-id: file:///var/svn/phpvms/trunk@869 e188d3d8-5a48-0410-b2de-ed2772025908

+++ install/install.sql
@@ -153,8 +153,8 @@ CREATE TABLE IF NOT EXISTS `phpvms_navdb` (
   `name` varchar(5) NOT NULL,
   `title` varchar(10) NOT NULL,
   `freq` varchar(7) NOT NULL,
-  `lat` float NOT NULL,
-  `lng` float NOT NULL,
+  `lat` float (8,6) NOT NULL,
+  `lng` float (9,6) NOT NULL,
   PRIMARY KEY (`id`),
   KEY `name` (`name`)
 ) ENGINE=MyISAM;
